LayoutTests/platform/gtk/http/tests/w3c/webperf/approved/navigation-timing/html5/test_timing_xserver_redirect-expected.txt

Description

This test validates the values of the window.performance.redirectCount and the window.performance.timing.redirectStart/End times for a cross-origin server side redirect navigation.




FAIL Starting document.location.hostname is correct (127.0.0.1:8000) assert_equals: Starting document.location.hostname is correct (127.0.0.1:8000) expected "127.0.0.1:8000" but got "127.0.0.1"(stack: assert@http://127.0.0.1:8000/w3c/resources/testharness.js:1646
assert_equals@http://127.0.0.1:8000/w3c/resources/testharness.js:567
http://127.0.0.1:8000/w3c/webperf/resources/webperftestharness.js:145
step@http://127.0.0.1:8000/w3c/resources/testharness.js:876
test@http://127.0.0.1:8000/w3c/resources/testharness.js:338
wp_test@http://127.0.0.1:8000/w3c/webperf/resources/webperftestharness.js:67
test_equals@http://127.0.0.1:8000/w3c/webperf/resources/webperftestharness.js:145
global code@http://127.0.0.1:8000/w3c/webperf/approved/navigation-timing/html5/test_timing_xserver_redirect.html:78)
PASS window.performance is defined 
PASS window.performance.navigation is defined 
PASS timing.navigation.type is TYPE_NAVIGATE 
PASS navigation.redirectCount == 0 on a cross-origin server redirected navigation 
PASS window.performance.timing.navigationStart > 0 
PASS timing.redirectStart == 0 on a server redirected navigation from another domain 
PASS timing.redirectEnd == 0 on a server redirected navigation from another domain 
